docs(snarea): Stage 3 runbook/architecture + minor fixes (docstrings, default_curve validation)
